move logger.py in utils directory
