+++ hydrabus_framework/core/engine.py
@@ -7,7 +7,6 @@ from importlib import import_module
 from prompt_toolkit.shortcuts import PromptSession
 from prompt_toolkit.styles import Style
 
-from hydrabus_framework.core.logger import Logger
 from hydrabus_framework.core.command.run import run_module
 from hydrabus_framework.core.command.show import show
 from hydrabus_framework.core.command.set_options import set_options
@@ -16,6 +15,7 @@ from hydrabus_framework.core.command.back import back
 from hydrabus_framework.core.command.quit import hbf_exit
 from hydrabus_framework.core.command.help import hbf_help
 from hydrabus_framework.modules.base import ABaseModule
+from hydrabus_framework.utils.logger import Logger
 
 
 __author__ = "Jordan Ovrè <ghecko78@gmail.com>"
