+++ hydrabus_framework/utils/protocols/uart.py
@@ -1,4 +1,4 @@
-from hydrabus_framework.core.logger import Logger
+from hydrabus_framework.utils.logger import Logger
 from hydrabus_framework.utils.generic import *
